# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API DOCUMENTATION",
        "title": "API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "prastiyo.beka12@gmail.com"
        }
    },
    "host": "localhost:3000",
    "paths": {
...
        "/api/v1/campaign": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Campaign",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Campaign Create",
                "parameters": [
                    {
                        "description": "Campaign Create",
                        "name": "campaign",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Campaign.CampaignRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/campaign/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Campaign Delete",
                "parameters": [
                    {
                        "description": "Campaign Update",
                        "name": "campaign",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Campaign.CampaignUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/campaign/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Campaign By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Campaign By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Campaign Update",
                "parameters": [
                    {
                        "description": "Campaign Update",
                        "name": "campaign",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Campaign.CampaignUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/dapil": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Dapil",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dapil"
                ],
                "summary": "Dapil",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Dapil",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dapil"
                ],
                "summary": "Dapil Create",
                "parameters": [
                    {
                        "description": "Dapil Create",
                        "name": "dapil",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Dapil.DapilRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/dapil/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Dapil",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dapil"
                ],
                "summary": "Dapil Delete",
                "parameters": [
                    {
                        "description": "Dapil Update",
                        "name": "dapil",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Dapil.DapilUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Dapil ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/dapil/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Dapil By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dapil"
                ],
                "summary": "Dapil By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dapil ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Dapil",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dapil"
                ],
                "summary": "Dapil Update",
                "parameters": [
                    {
                        "description": "Dapil Update",
                        "name": "dapil",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Dapil.DapilUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Dapil ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data profile user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Credential"
                ],
                "summary": "Get User",
                "responses": {}
            }
        },
...
        "/api/v1/user": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Create",
                "parameters": [
                    {
                        "description": "User Create",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User.UserRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/user/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Delete",
                "parameters": [
                    {
                        "description": "User Update",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User.UserUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data User By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Update",
                "parameters": [
                    {
                        "description": "User Update",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User.UserUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "Campaign.CampaignRequest": {
            "type": "object",
            "required": [
                "campaign",
                "company_id",
                "profile_id"
            ],
            "properties": {
                "campaign": {
                    "type": "string"
                },
                "campaign_amount": {
                    "type": "integer"
                },
                "campaign_status": {
                    "type": "integer"
                },
                "campaign_target": {
                    "type": "integer"
                },
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "profile_id": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "tps_id": {
                    "type": "string"
                }
            }
        },
        "Campaign.CampaignUpdateRequest": {
            "type": "object",
            "properties": {
                "campaign": {
                    "type": "string"
                },
                "campaign_amount": {
                    "type": "integer"
                },
                "campaign_status": {
                    "type": "integer"
                },
                "campaign_target": {
                    "type": "integer"
                },
                "company_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "profile_id": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "tps_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "Dapil.DapilRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "Dapil.DapilUpdateRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "User.UserRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "dapil_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rw": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "village_id": {
                    "type": "string"
                }
            }
        },
        "User.UserUpdateRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "dapil_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rw": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "village_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}